requirements.txt
loguru~=0.5.3
kafka-python~=2.0.2
pydantic~=2.5.3